in profile list of animals you have, fixes in database(-table user_pet

+++ FinalTask/src/main/java/by/sazanchuk/finalTask/service/PetServiceImpl.java
@@ -2,11 +2,16 @@ package by.sazanchuk.finalTask.service;
 
 import by.sazanchuk.finalTask.dao.DaoException;
 import by.sazanchuk.finalTask.dao.PetDao;
+import by.sazanchuk.finalTask.dao.connectionPool.ConnectionPoolException;
 import by.sazanchuk.finalTask.entity.Pet;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class PetServiceImpl extends ServiceImpl implements PetService{
+    PetServiceImpl() throws DaoException, ConnectionPoolException {
+    }
+
     @Override
     public List<Pet> findAll() throws DaoException {
         PetDao petDao = transaction.createDao(PetDao.class);
@@ -24,6 +29,7 @@ public class PetServiceImpl extends ServiceImpl implements PetService{
         PetDao petDao = transaction.createDao(PetDao.class);
         if (pet.getIdentity() == null) {
             pet.setIdentity(petDao.create(pet));
+//            petDao.createPetUser(pet);
         }
         petDao.update(pet);
         return pet.getIdentity();
@@ -32,12 +38,21 @@ public class PetServiceImpl extends ServiceImpl implements PetService{
     @Override
     public void delete(Integer identity) throws DaoException {
         PetDao petDao = transaction.createDao(PetDao.class);
-        petDao.delete(identity);
+        if (identity != null) {
+            petDao.delete(identity);
+        }
     }
 
     public List<Pet> getPetsOfOneUser(Integer userId) throws DaoException {
         PetDao petDao = transaction.createDao(PetDao.class);
-        return petDao.readPetsWithOneUser(userId);
+        List<Pet> pets = petDao.readPetsWithOneUser(userId);
+        List<Pet> petList = new ArrayList<>();
+        for (Pet pet: pets){
+            Pet p = new Pet();
+            p = petDao.read(pet.getIdentity());
+            petList.add(p);
+        }
+        return petList;
     }
 
     public Pet findByNameAndUserId(String name, Integer user_id) throws DaoException {
